pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Mattheo Mustert mattheo.mustert@simonloos.nl"]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.3"
click = "^8.1.7"
pyarrow = "^14.0.1"
python-dotenv = "^1.0.0"
google-cloud-storage = "^2.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
